price calculation

+++ club_service/domain/price_period.go
@@ -3,26 +3,26 @@ package domain
 // PricePeriod use for configuration price in minutes.
 // f.e. for PricePeriod from 0 minute to 60 minute Price will be 10$ per minute
 type PricePeriod struct {
-	from int64
-	to   int64
+	from int
+	to   int
 }
 
-func NewPricePeriod(from int64, to int64) *PricePeriod {
+func NewPricePeriod(from int, to int) *PricePeriod {
 	return &PricePeriod{
 		from: from,
 		to:   to,
 	}
 }
 
-func (p PricePeriod) From() int64 {
+func (p PricePeriod) From() int {
 	return p.from
 }
 
-func (p PricePeriod) To() int64 {
+func (p PricePeriod) To() int {
 	return p.to
 }
 
-func (p *PricePeriod) TimeForPay(durationMinutes int64) int64 {
+func (p *PricePeriod) TimeForPay(durationMinutes int) int {
 	if p.from > durationMinutes {
 		return 0
 	}
@@ -34,6 +34,6 @@ func (p *PricePeriod) TimeForPay(durationMinutes int64) int64 {
 	return p.totalTime()
 }
 
-func (p PricePeriod) totalTime() int64 {
+func (p PricePeriod) totalTime() int {
 	return p.to - p.from
 }
